+++ CodexLib/Classes/PartCustomData.cs
@@ -32,6 +32,7 @@ namespace CodexLib
             unit.Parts.RemoveAll<PartCustomData>(f => true);
         }
 
+        [Obsolete]
         public static T GetData<T>(this UnitEntityData unit, string key) where T : class
         {
             var data = unit?.Get<PartCustomData>();
@@ -41,12 +42,14 @@ namespace CodexLib
             return value as T;
         }
 
+        [Obsolete]
         public static void AddData(this UnitEntityData unit, string key, object value)
         {
             var data = unit?.Ensure<PartCustomData>();
             data.Data[key] = value;
         }
 
+        [Obsolete]
         public static void RemoveData(this UnitEntityData unit, string key)
         {
             var data = unit?.Get<PartCustomData>();