+++ DarkCodex/Main.cs
@@ -499,6 +499,7 @@ namespace Shared
             LoadSafe(Kineticist.CreateExpandedElement);
             LoadSafe(Kineticist.PatchVarious);
             LoadSafe(Kineticist.FixBlastsAreSpellLike);
+            LoadSafe(Kineticist.CreateVenomInfusion); // keep late
             LoadSafe(Kineticist.CreateSelectiveMetakinesis); // keep late
 
             // Monk