+++ DarkCodex/changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 - ~~added save metadata to keep track of enabled patches~~ not yet
 
+## [1.3.13]
 - fixed sticky spells not working with Boundless Injury
 - allowed arcanist to use temporary known spells without preparing them
+- added Venom Infusion; infusion: applies sickened or poisons the target
 
 ## [1.3.12]
 - added Patch Fix Arcanist Spontaneous Metamagic: allows arcanist to use non memorized metamagic, but increases casting time